+++ app/src/main/java/com/architectcoders/aacboard/di/CommonModule.kt
@@ -1,9 +1,9 @@
 package com.architectcoders.aacboard.di
 
-import com.architectcoders.aacboard.data.repository.PermissionChecker
-import com.architectcoders.aacboard.datasource.PermissionCheckerImpl
+import com.architectcoders.aacboard.data.datasource.local.AppPermissionChecker
+import com.architectcoders.aacboard.datasource.AppPermissionCheckerImpl
 import org.koin.dsl.module
 
 val commonModule = module {
-    factory<PermissionChecker> { PermissionCheckerImpl(get()) }
+    factory<AppPermissionChecker> { AppPermissionCheckerImpl(get()) }
 }
