+++ data/src/main/java/com/architectcoders/aacboard/data/datasource/local/AppPermissionChecker.kt
@@ -1,0 +1,8 @@
+package com.architectcoders.aacboard.data.datasource.local
+
+interface AppPermissionChecker {
+
+    enum class Permission { COARSE_LOCATION }
+
+    suspend fun check(permission: Permission): Boolean
+}
