+++ data/src/main/java/com/architectcoders/aacboard/data/datasource/local/DeviceDataSource.kt
@@ -5,5 +5,4 @@ import kotlinx.coroutines.flow.Flow
 interface DeviceDataSource {
     fun getPreferredDashboardId(): Flow<Int>
     suspend fun setPreferredDashboardId(id: Int)
-    suspend fun getLastUserRegion():String?
 }
