+++ data/src/main/java/com/architectcoders/aacboard/data/datasource/local/LocationDataSource.kt
@@ -1,0 +1,5 @@
+package com.architectcoders.aacboard.data.datasource.local
+
+interface LocationDataSource {
+    suspend fun getLastUserRegion():String?
+}
