+++ data/src/main/java/com/architectcoders/aacboard/data/repository/RegionRepositoryImpl.kt
@@ -1,30 +1,23 @@
 package com.architectcoders.aacboard.data.repository
 
-import com.architectcoders.aacboard.data.datasource.local.DeviceDataSource
+import com.architectcoders.aacboard.data.datasource.local.LocationDataSource
+import com.architectcoders.aacboard.data.datasource.local.AppPermissionChecker
 import com.architectcoders.aacboard.domain.repository.RegionRepository
 
 class RegionRepositoryImpl(
-    private val deviceDataSource: DeviceDataSource,
-    private val permissionChecker: PermissionChecker
+    private val locationDataSource: LocationDataSource,
+    private val appPermissionChecker: AppPermissionChecker
 ) : RegionRepository {
 
     override suspend fun getLastUserRegion(): String {
-        return if (permissionChecker.check(PermissionChecker.Permission.COARSE_LOCATION)) {
-            deviceDataSource.getLastUserRegion()?.lowercase() ?: DEFAULT_USER_REGION
+        return if (appPermissionChecker.check(AppPermissionChecker.Permission.COARSE_LOCATION)) {
+            locationDataSource.getLastUserRegion()?.lowercase() ?: DEFAULT_USER_REGION
         } else {
             DEFAULT_USER_REGION
         }
     }
 
-
     companion object {
         private const val DEFAULT_USER_REGION = "us"
     }
-}
-
-interface PermissionChecker {
-
-    enum class Permission { COARSE_LOCATION }
-
-    suspend fun check(permission: Permission): Boolean
 }
